pim to core 3.22.x
